Reland "Revert "hide nested hinting enum""
This reverts commit f49563bf9276f70441fd22da83413d585b2bcc17.

Reason for revert: Google3 also.

Original change's description:
> Revert "Revert "hide nested hinting enum""
> 
> This reverts commit 6bd19df9fa9d26d9de4de646b2cdf82ff5dca67c.
> 
> Restores original CL, but adds guards for flutter.
> 
> Bug: skia:
> Reviewed-on: https://skia-review.googlesource.com/c/169345
> Commit-Queue: Mike Reed <reed@google.com>
> Auto-Submit: Mike Reed <reed@google.com>

TBR=fmalita@chromium.org,reed@google.com

No-Presubmit: true
No-Tree-Checks: true
No-Try: true
Bug: skia:

+++ src/utils/SkLua.cpp
@@ -818,7 +818,7 @@ static int lpaint_setTypeface(lua_State* L) {
 }
 
 static int lpaint_getHinting(lua_State* L) {
-    SkLua(L).pushU32((unsigned)get_obj<SkPaint>(L, 1)->getHinting());
+    SkLua(L).pushU32(get_obj<SkPaint>(L, 1)->getHinting());
     return 1;
 }
